Add in-cluster support
