+++ README.md
@@ -3,7 +3,7 @@
 
 ----
 
-logga is an easy to use and powerful container log dashboard for Kubernetes. It's used by application teams and sysadmins to effortlessly view container logs in a namespace using a web browser, without a complex logging infrastructure. Logga is ment to be deployed in Kubernetes per namespace by developers and uses in-cluster RBAC authorisation. Some of the key futures are:
+logga is an easy to use and powerful container log dashboard for [Kubernetes](https://kubernetes.io). It's used by application teams, developers and sysadmins to effortlessly view container logs in a namespace using a web browser, without a complex logging infrastructure. Logga is ment to be deployed in Kubernetes per namespace by developers and uses in-cluster RBAC authorisation. Some of the key futures are:
 
 * Super easy do deploy and manage
 * Fully stateless 
@@ -29,4 +29,4 @@ docker run -d --name logga -p 8080:8080 amimof/logga:latest
 
 ## Contributing
 
-All help in any form is highly appreciated and your are welcome participate in developing `Huego` together. To contribute submit a `Pull Request`. If you want to provide feedback, open up a Github `Issue` or contact me personally. 
+All help in any form is highly appreciated and your are welcome participate in developing `logga` together. To contribute submit a `Pull Request`. If you want to provide feedback, open up a Github `Issue` or contact me personally.
