+++ pkg/api/broker.go
@@ -1,0 +1,64 @@
+package api
+
+import (
+	"log"
+	"time"
+)
+
+type Broker struct {
+	// Events are pushed to this channel by the main events-gathering routine
+	Notifier chan []byte
+
+	// New client connections
+	newClients chan chan []byte
+
+	// Closed client connections
+	closingClients chan chan []byte
+
+	// Client connections registry
+	clients map[chan []byte]bool
+}
+
+func NewBroker() *Broker {
+	broker := &Broker{
+		Notifier:       make(chan []byte, 1),
+		newClients:     make(chan chan []byte),
+		closingClients: make(chan chan []byte),
+		clients:        make(map[chan []byte]bool),
+	}
+	go broker.listen()
+	return broker
+}
+
+func (broker *Broker) listen() {
+	patience := time.Second * 1
+	for {
+		select {
+		case s := <-broker.newClients:
+
+			// A new client has connected.
+			// Register their message channel
+			broker.clients[s] = true
+			log.Printf("Client added. %d registered clients", len(broker.clients))
+		case s := <-broker.closingClients:
+
+			// A client has dettached and we want to
+			// stop sending them messages.
+			delete(broker.clients, s)
+			log.Printf("Removed client. %d registered clients", len(broker.clients))
+		case event := <-broker.Notifier:
+
+			// We got a new event from the outside!
+			// Send event to all connected clients
+			for clientMessageChan, _ := range broker.clients {
+				select {
+				case clientMessageChan <- event:
+				case <-time.After(patience):
+					//log.Print("Skipping client.")
+					continue
+				}
+			}
+		}
+	}
+
+}
